Include coordinates when fetching locations (#20)
* feat: scrape and include coordinates in location fetch (#18)

Coordinates of parishes are now included by default when scraping locations. The new fields `longitude` and `latitude` contain floats.

Because this effectively doubles the amount of requests, be aware that the extra information comes with a price. This feature can be disabled by using the flag `--exclude-coordinates`. Use it if you suffer decreased performance.