feat: scrape church registers, i.e. images (#3)
This commit implements a basic version allowing users to scrape
Matricula's most important resource: scanned images of church registers.
With `matricula-online-scraper get church-register <URL|stdin>` can the images
of one or more church registers be scraped. Use `--help` for that specific
commit for details.

Note that the terminology wasn't fixed. Instead of "church registers", previous
code might refer to this as "parish registers". From now on, the former
should be used. This is the term Matricula uses itself.

The spider is already grouped within the new command `get`, while others are
still grouped in `fetch`. This conforms with a future refactoring,
enhancing the CLI structure.

https://github.com/lsg551/matricula-online-scraper/issues/3#issuecomment-2498586139
was not implemented directly as a CLI option, because the command can handle
multipel URLs, either per argument or read from stdin.
In a future refactoring the other commands should be refined so that chainig
is possible: `command1 | command2`. This allows to pipe a list of all
church registers from e.g. one specific parish into the image scraper.